+++ templates/definition/meter/homeassistant.yaml
@@ -9,6 +9,9 @@ requirements:
   description:
     en: Requires a running Home Assistant instance with suitable sensor entities. All values are Home Assistant entity IDs (e.g. sensor.*)
     de: Erfordert eine laufende Home Assistant Instanz mit passenden Sensor-Entitäten. Alle Werte sind Home Assistant Entity IDs (z.B. sensor.*)
+auth:
+  type: homeassistant
+  params: [home]
 params:
   - name: usage
     choice: ["grid", "pv", "battery", "aux", "charge"]
