+++ templates/definition/vehicle/homeassistant.yaml
@@ -9,6 +9,9 @@ requirements:
   description:
     en: Requires a running Home Assistant instance with suitable vehicle entities and services. All values are Home Assistant entity IDs (e.g. sensor.*, binary_sensor.*, number.*, script.*)
     de: Erfordert eine laufende Home Assistant Instanz mit passenden Fahrzeug-Entities und Services. Alle Werte sind Home Assistant Entity IDs (z.B. sensor.*, binary_sensor.*, number.*, script.*)
+auth:
+  type: homeassistant
+  params: [home]
 params:
   - preset: vehicle-common
   - name: uri
